Update to 1.3.0
1.3.0
  * 一定期間経過した新着情報を削除する機能を実装しました。
  * 所属選択UIにおいて空白が選択された場合（何も選択されていない状態）の処理を追加しました。
  * 回覧板の添付ファイル出力が出来ない問題を修正しました。
  * 権限がないユーザーにもスケジュール画面でプラスボタンが出る場合がある問題を修正しました。
  * 施設マスタの利用開始日、終了日のエラーチェックを一部修正しました。
  * ファイル管理のフォルダ作成・編集時、グループ閲覧権限に「制限なし」項目を追加しました。
  * グループの上位階層を選択できるよう修正しました。
  * 上位グループが異なる場合同一のグループコードを持つグループを登録できる問題を修正しました。
  * 所属情報を持たないユーザーもユーザー一覧に表示するよう修正しました。
  * ユーザーの所属が削除できない場合がある問題を修正しました。
  * ページのタイトルを「縁sys市」から「縁sys」に変更しました。

+++ lib/gw/script/reminder.rb
@@ -1,11 +1,20 @@
 class Gw::Script::Reminder
   class << self
-    def clear
+    def clear(opt = {seen: true, old: nil})
       log_write("Reminder.clear start")
-      reminders = Gw::Reminder.unscoped.where(
-                     Gw::Reminder.arel_table[:seen_at].not_eq(nil))
-      delcnt = reminders.destroy_all.size
-      log_write("#{delcnt} reminders deleted.")
+      if opt[:seen]
+        reminders = Gw::Reminder.unscoped.where(
+                       Gw::Reminder.arel_table[:seen_at].not_eq(nil))
+        delcnt = reminders.destroy_all.size
+        log_write("#{delcnt} seen reminders deleted.")
+      end
+      if opt[:old]
+        reminders =
+          Gw::Reminder.unscoped.where(
+            Gw::Reminder.arel_table[:updated_at].lt(eval(opt[:old]).ago))
+        delcnt = reminders.destroy_all.size
+        log_write("#{delcnt} old reminders deleted.")
+      end
       log_write("Reminder.clear end")
     end
